add docs && small fixes

+++ fd/nt_putchar_fd.c
@@ -1,6 +1,12 @@
 # include "nt_fd.h"
 
-int nt_putchar_fd(const char c, const int fd)
+/**
+ * Write a character in a file
+ * @param c - The character
+ * @param fd - The file descriptor
+ * @return - The number of byte written (-1 if there was an error)
+ */
+ssize_t nt_putchar_fd(const char c, const int fd)
 {
     return write(fd, &c, 1);
 }
