+++ fd/nt_putnbr_fd.c
@@ -1,6 +1,12 @@
 # include "nt_fd.h"
 
-int nt_putnbr_fd(int n, const int fd)
+/**
+ * Write a number in a file
+ * @param n - The number
+ * @param fd - The file descriptor
+ * @return The number of bytes written (-1 if there was an error)
+ */
+ssize_t nt_putnbr_fd(int n, const int fd)
 {
     if (n <= MIN_INT)
     {
